Add improved tools for parsing through emails
Improves data retrieval from Gmails. Gmails contain both text
and html parts, hence why there are multiple parts. These scripts
prioritize the html, which are parsed with BeautifulSoup. This
removes the links, but can add in excessive \n,\t, or \r. Some of
this is filtered, others are left in to later replace with
whitespace. More can be done to improve text for the other tracks.