time extraction
extracts time from a csv file containing a single row. row consists of 3 columns. column 2 must contain the full string with the body of the email.

extractTime method returns type datetime. it returns None if date written is considered invalid.

an invalid date is a date with the year<current year, or a  date that only contains month with no date.

a date is valid as long as it has time or month and date, or both.

if the email only contains a time, it assumes the time is today.